requirements_pip.txt
pytorch_lightning==1.6.5
timm==0.6.11
einops==0.5.0
pandas==1.3.5
albumentations==1.3.0
pytorch-wavelet==1.2
torchmetrics=0.10.0